Tiny, strange fix for Pixmap mangling.
The right-half icon was getting drawn or filtered badly before, with the edges of lines replaced by repeating short gradients. The fix turned out to be to use more of the input Pixmap, actually sampling a large area that is out of bounds, and to write that to a larger output area as well. I'm not sure why it works, but it does.